requirements.txt
aiofiles            23.2.1
aiogram             3.13.0
aiohappyeyeballs    2.4.0
aiohttp             3.10.5
aiosignal           1.3.1
annotated-types     0.7.0
anyio               4.6.2.post1
attrs               24.2.0
certifi             2024.8.30
charset-normalizer  3.3.2
frozenlist          1.4.1
h11                 0.14.0
httpcore            1.0.6
httpx               0.27.2
idna                3.8
magic-filter        1.0.12
multidict           6.1.0
pip                 24.3.1
pydantic            2.8.2
pydantic_core       2.20.1
pyTelegramBotAPI    4.22.1
python-dotenv       1.0.1
python-telegram-bot 21.6
requests            2.32.3
sniffio             1.3.1
typing_extensions   4.12.2
urllib3             2.2.2
yarl                1.11.1
